Add demo data initialization on table creation

+++ src/main/java/org/alexej/demo/repository/ProductRepository.java
@@ -2,4 +2,4 @@ package org.alexej.demo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {}
+public interface ProductRepository extends JpaRepository<Product, Long> {}
